Renaming biggestAvailableArray to biggestAvailableArea + removing duplicated code.
